Implemented the base NPC class
I don't think we'll ever even use NPCs, but it was a really helpful exercise to better understand the modularity and structure of our implementation of the GAS framework. Maybe it's a bad sign that I needed an exercise to understand a system I came up with, but GAS is such a complex framework that I'll give myself a pass.