Implemented jumping, falling, and landing backend and animation logic
I might be the smartest human alive.

This required a bit of refactoring, but it worked out really well. I created a new a base character class that has broadcasts a delegate when the character jumps and implements the custom character movement component.

Added new delegates to the custom character movement component that adds a "jumping" tag to the character when they jump and a "falling" tag when they begin falling. The tags are all cleared when the character lands.

The new animation logic functions off of gameplay tag logic instead of constantly checking movement states and velocity.