Created the TriggerGameplayEvent animation notify
There is a commonly implemented animation notify that sends a gameplay event to an actor at a certain time during an animation. This is used for ability logic that occurs at a specific point in their animation. E.g. a fireball should not be spawned until after the windup section of its animation.