Implemented the "Death" ability, integrated with the GM
Created a new "Death" gameplay ability that is automatically triggered on player death. The ability cancels other ongoing abilities and calls a "StartDeath" function in the game mode. The ability then sets a timer for "DeathDuration," after which it ends itself, triggering a "FinishDeath" function in the game mode. Most of the death logic will be implemented in the "StartDeath" and "FinishDeath" functions of the game mode.