Implemented dynamic ability data
Created a new "DynamicAbilityData" class, which will be used to manage dynamic ability data, such as which actors should be spawned depending on the player's equipped skin.

Also implemented a new "DeathEvent" delegate that is fired from the ASC when the Death event is received. This allows the DeathData to be passed to anyone who needs it, such as the dying character. I added death-launch mechanics to the dying ragdoll to showcase this, using the killing-damage magnitude, retrieved from the death data, as a scaler for the force with which the ragdoll is launched.