Fixed FPP "Jump" state machine bug
Fixed a bug where first-person challenger animation instances sometimes would not enter the "Jump" state machine. This was occurring when playing on a listen server, where the ASC was being initialized before the animation instance had a change to bind to its OnInitialized delegate, which prevented the animation instance from ever caching the initialized ASC. I fixed this by checking if the ASC is already initialized when the animation instance is being initialized and manually calling OnASCInitialized if it is.

All game systems now work on both listen servers and dedicated servers again.