Reorganized environment asset folder structure
Created two new folders for environment assets: "EditorOnlyAssets" and "Environment." EditorOnlyAssets contains every asset included in the modular asset packs used for this project. When we put an asset into a level for the first time, we move it to the Environment folder. The EditorOnlyAssets folder is not cooked or included in packaged builds.

Since we're using pre-made asset packs, we have a lot of assets that we don't use. This method of organizing assets effectively excludes unused assets from our builds. This one change reduced the game's build size by 50% (down to about 300mb from over 600mb).