[v0.1.2] Implemented asset manager utilities, refactored game framework, and implemented the main menu
Implemented utilities for loading global game data in the asset manager. For example, when in the main menu, we'll want to have the main menu's UI data, which is static, and will never change during runtime. Instead of trying to put it into the game mode or player controller, we can use the asset manager to load it anywhere we need it, when we need it, and unload it when we're done. This is significantly more memory-efficient and keeps our code much cleaner and more consistent. The asset manager framework I've implemented here will also serve as the foundation for the asset manager's future responsibilities, such as loading and unloading game data during runtime, between levels.

Refactored the game mode class hierarchy. I've decided to commit to the ACrashGameMode, a child of the AGameMode, being the base class for the wide majority of game modes. Other game modes, like the main menu game mode, will simply be removed from this project's hierarchy, deriving from AGameMode or AGameModeBase directly. This is a classic inheritance vs. composition problem, since we're using two different built-in Unreal Engine classes as the bases for our classes. This would be avoided if we implemented our own ACrashGameMode from AGameModeBase, which we may do in the future.

Lastly, implemented the main menu system. Standalone players, clients, and listen servers in the editor or in a packaged build now load into the main menu level and are given a modular main menu user interface. They can then travel to game levels and play the game with the correct game framework classes instead of the main menu's classes. For testing, I've just been using commands for travelling. But soon, I'll implement the frontend of the main menu.

Packaged v0.1.2 with these changes.