Installed GameplayMessageRouter plugin and set up hit-markers
Installed Epic's Gameplay Message Router plugin. This is literally one of the most useful plugins in the engine and it's hidden in the middle of Lyra.

Used the plugin to implement hit-markers. The hit-marker animation will now play each time the local player does damage. This is REALLY useful for play-testing (and, obviously, for players).