Debugged hit-markers and life counter
I realized that the message router was not replicated by default, so I created a helper function in the game state that can replicate messages to clients, which allowed hit-markers and elimination-markers to be replicated to clients.

The life counter bug was caused by some strange order-of-execution problem where clients connecting to listen servers did not have their player states initialized before they created their UI. To fix this, I bound a callback to when the UI's owning player controller's player state changes, which then performs the necessary logic using that new player state.