Created the framework for the character selection system
Implemented the framework for the character selection system. The character selection widget will now be created when players join the game in a "waiting to start match" state; the character selection background level will also be streamed in at this time.

Created the Knight's character selection animations, animation blueprint, and actors.

Most of these changed files are a result of slightly restructuring the folder hierarchy, and making small changes to the project's base skeleton.