Refactored equipment system to be server-authoritative
I initially designed this system to be server-authoritative, but I ended up implementing it on the client side. I refactored it to only allow equipment to be equipped and unequipped via the server, and for it to then be replicated to clients.

This system should be server-authoritative because it's in charge of granting abilities and effects. This system would also have major ramifications if the client and server were to become desynced.