Debugged challenger and ability selection bugs
I'm still not sure what caused this bug or how I fixed it, but I'm so done with this...

This bug caused certain ability buttons to be selected when created for no apparent reason. Specifically, it always selected the Knight's "Parry" or "Warhammer" ability when switching to the Knight from any other Challenger.