Added temporary fix for action widget initialization
Added a temporary fix for the initialization of action widgets. The Common UI Subsystem's "input method changed" delegate needs to be broadcast for the icons to appear properly, but something isn't being initialized in time, so I just set an arbitrary timer to broadcast it after a few seconds to ensure everything is initialized first. I'll replace this when I figure out what's causing the initialization to fail.d

Also styled the action widget icon itself.