Fixed movement tag and ability slot bugs
Fixed a bug that prevented movement tags from being cleared when a player dies, causing them to respawn with the tags with which they died (e.g. "falling," "jumping", etc.).

Also fixed a bug that caused ability slot widgets to get stuck in an "active" state when an ability is ended too quickly after being activated.

Added preconstruction code to only display the border for ability input icons for MnK icons, hiding it for gamepad icons.