Implemented skin-based ability cues
Skins can now define custom gameplay cues. When an ability wants to activate a gameplay cue, it will now check for a matching cue in the skin, which it will use instead. This allows skins to easily define custom VFX, sound effects, and decals that will be used in abilities when the skin is equipped.