Debugged equipment prediction rollback
Cleaning equipment prediction data too quickly (e.g. when the task ends) meant properly-predicted equipment would get rolled back when the client eventually caught up to the server, since the prediction appeared to be null at that point. The prediction is essentially racing against the task itself.

I made some tweaks that allows us to hold off on cleansing prediction data until a NEW prediction occurs. This gives ensures that if a prediction is correct, it will always be confirmed by the server eventually, unless a new prediction needs to happen, in which case, the old prediction is irrelevant.