Moved ability parameters into each ability's properties
Removed the "static ability data" asset, which was used to centralize data used for abilities (damage, range, etc.). Moved each ability's data into its own properties. This is kind of the whole point of GAS: that each ability is self-contained. It was pointless to abstract all of that data into one asset that had to be edited (and loaded) separately.