Integrated base game state and game mode with the initialization state framework
Did some initial testing to see if we can use the initialization state framework (IGameFrameworkInitStateInterface) for initializing our game framework, and successfully implemented initialization states into the game state and game mode, and tied them together. The game mode now requires the game state to be in "Initializing" state before it can transition to the "Initializing" state. The game state notifies the game mode when it transitions.