Implemented developer settings
Created new developer settings that will appear in the editor settings. These can be used to test various game features that typically require game options or game session settings.

For example, instead of having to go through the character selection screen to choose a Challenger, the desired Challenger can be overridden in the developer settings, and the game will use that Challenger instead of searching for one from the game options.