Completed game mode initialization implementation
Also created a new "world settings" class that defines level-specific properties, like the intro cinematic to play, or the default game mode data to use (specifically for game modes without another means of finding GM data, like the main menu).

Also implemented the intro cinematic into the player controller, which now plays the cinematic in a loop when the game starts, if desired by world settings. This will be stopped when players are ready to be initialized.