Implemented game mode loading via game state and created "WinterHolidaysEvent" feature
Implemented the game state's game mode loading logic. This is responsible for loading the game mode data, its appropriate bundles, and its game features, as well as activating its actions.

Created a new "WinterHolidaysEvent" game feature to test the game feature system. This game feature represents a hypothetical limited-time event or add-on with its own characters, abilities, maps, etc. When the plugin is enabled, all of its assets are cooked into the game and registered by the asset manager, allowing content to be added with full modularity.