Created new ProjectCrashEditor module
This is a new editor module that will be used to create custom editor features. It will first be used to dynamically activate and deactivate game features during PIE.

By default, the engine loads assets as they are needed; but in the editor, it never unloads them (by design). We want override this behavior when it comes to game features, to simulate their behavior in packaged builds, where they are dynamically loaded, activated, deactivated, and unloaded as needed.