Changed to delegate for listening to game mode load callbacks from game state
The game mode manager component previously called CheckDefaultInitialization directly on its owning GameState when it finished loading its game mode, since the game state is dependent on the game mode being loaded to continue initialization. Now, the game state instead listens for a delegate in the game mode manager that is broadcast when the game mode is loaded.

These two classes should always be together, so this doesn't matter all too much. It's just good to decouple their logic.