Implemented a CharacterWithAbilities class
This is a subclass of CrashCharacter, that also uses the pawn extension component. The difference is that this class contains its own ASC, instead of using one from a player state. The pawn extension component is now able to recognize this, and can still properly initialize with the character's ASC.