Created the CameraModeStack class
This class will be used by the CrashCameraComponent to manage switching between multiple camera modes.

For example, an ability may want to change the camera (think of Reinhardt's shield, supers in Destiny, etc.). It can push its desired camera mode to this stack, the camera will blend to it, then when the ability ends, it can pop the camera mode from the stack, and the camera will return to its previous camera.

This is very similar to the ChallengerBase class's camera stack system. However, the Challenger only came with two camera modes: first-person and third-person, and did not support modular addition of new cameras.