Completed base first-person character animation instance
This class, UFirstPersonCharacterAnimInstance, performs movement sway and aim sway calculations using a spring model defined by the animation instance.

It's difficult to test whether this works without an actual animations, but I've been working with this kind of animation system for so long that I can tell it's functioning properly just by reading the spring values.