Began moving (currently) unused assets from editor-only asset folder to cooked content folders
It's a huge mess having two content folders: one for assets currently being used in levels, and one for uncooked, unused assets. It messes with the asset manager and data validation, and it slows down the level design process.

I've made the decision to cook all assets, even if they aren't used. This will increase the packaged build size, but I think it's worth it for how much easier it will make development. Additionally, the tradeoff will balance out eventually, as we'll eventually end up using the majority of these assets anyway.

This is going to be a pretty long process (there's about 5,000 assets to move). I've only finished moving the PolygonAdventure folder.