Adjustments to first-person character animation class for edge cases
Replaced the rotation delta calculation with a normalized delta to account for wrapping. For example, when the player's rotation changes from 359° to 2° in one frame, we should assume that the change was 3°, not -357°.

Also made minor optimizations: moved some internal constants to be macros as a minor optimization and applied spring interpolation scalars inline, instead of changing the spring variable twice.