Testing new offset-based first-person jump animations
Instead of creating manual transitions between jumping up, falling, reaching an apex, landing, etc., I've tried using an additional spring model to drive offsets based on vertical movement, and this looks SO much better, that I'm going to be replacing the previous system with this one.

In the next commit, I'll be removing the previous animation method, and formalizing the new one.