Implemented the "Reset to Bounds" ability
This ability will only be used in the "Practice Range." It listens for an "Out of Bounds" gameplay event (fired by a physics volume in the practice range) and resets the player to a safe location. A line trace is continuously performed while the player is grounded to update the location to which they should be reset.