Debugged modular control rig's foot controls
The default foot controls didn't play nicely with our extra "toes" bone. The outermost foot bone can now be controlled properly, but the foot ball and toes cannot be controlled with the modular control rig. I'm hoping to eventually find a solution to this, but we may just have to move on. These bones' transforms aren't really noticeable anyways.