Refactored input setup assets
Moved default playable character ("Challenger") input assets into their own independent data assets which are tied directly to their pawn data, rather than being reliant on the game mode for their ability input.

I think this structure makes more sense because pawn data is responsible for adding the pawn's abilities, so it should also be responsible for adding the input for those abilities. I believe it is easier and more manageable to create new input data and pawn data for special cases (e.g. Challengers that don't use a normal ability set, such as those used in a special game mode) than it is to be reliant on the game mode for that input data.