Implemented full-body third-person animation framework for forward/backward animations
The previous framework didn't work out. It was unintuitive to animate with and didn't get the visuals I wanted. I'm going to try a few different things, starting with just using full-body animations for each equipment set. This is looking really good so far, though I only have it working for forward/backward locomotion and leans.