TPP Animation System: Version B
This system re-uses a single set of global base animations for the root, legs, and spine of each equipment. Equipment defines its own arm poses, leans, and offsets.

Pros:
* Saves an enormous amount of work.
* Easy to animate (just simple poses).
* Looks great; no inertial blending issues.
* Animations are consistent from equipment-to-equipment and character-to-character.

Cons:
* Can't create unique arm animations for equipment.

I'm currently struggling to implement dynamic leans on a per-equipment basis without messing with the base animation. If I can figure this out, I think this will be the final animation system version we'll use. It saves an incredible amount of animation work *and* ensures all equipment animations are consistent, which I believe is desirable in a game like this, which emphasizes precision.