Updated inertialization methods for third-person animation
Dead blending inertialization is now performed twice for third-person animation. It is performed once for ground movement, accelerating blending of leg bones for smoother walk direction transitions.

Inertialization is then performed a second time in the animation graph. When grounded, inertialization has already been performed at this point in the "Walk" state, so it should have no effect. But when airborne, this is where inertialization for jumping transitions is performed.