Created the "Crash Team Agent" interface
This interface, CrashTeamAgentInterface, will be implemented by any objects that want to be on a team (player states, AI characters, etc.), and any objects that want to receive team events (player controllers, player avatars, etc.).

Implemented the CrashTeamAgentInterface on CrashPlayerState. The player state will be responsible for managing players' teams.