Created the "GameplayTagStack" structure
This structure represents a "stack" of gameplay tags. It contains a tag (the type of tag being stacked) and a count (the number of tags in the stack).

Tag stacks are managed by stack containers (GameplayTagStackContainer). These are a fast array serializer that can quickly and efficiently replicate tag stacks.

Tag stack containers will be used to track player and team statistics. E.g. the amount of damage that a player deals in a game can be represented by a stack of "DamageDealt" tags; a team's current score can be represented by a stack of "TeamScore" tags; etc.