Implemented "TeamInfo" actor
This class is a tangible representation of a team during gameplay. It is an actor, rather than an object, for replication purposes. Each team present in a game will have a TeamInfo actor created to represent it, which is responsible for tracking information like team statistics (e.g. team score).