Implemented team interface on CrashPlayerController
The CrashPlayerController can now be used to interface with the player's team. The player controller's team interface is driven by its player state.

This will be used to allow characters to drive their team logic with either a player state (via an owning player controller) or with an AI controller.