Debugged "Camera" bone controls
The "Camera" bone is now centered at (0,0,0), preventing it from influencing the camera when it's not being used by an animation. The "Camera Control" rig module now accounts for this, offsetting the control to apply in the correct space.

The camera bone is parented to the root, which isn't X-forward (that's Synty's fault), so the camera bone wouldn't properly align with the camera itself without this offset. I've made the offset a variable so we can use this for X-facing characters too.

There's still one more bug: the identifying camera tag returned by the camera stack isn't updating when changing between cameras.