Fixed issue with retrieving relevant camera mode info
Added a new GetBlendInfoBottom function to CrashCameraModeBase, which can be used to get the blend information of the camera which is being blended OUT of when transitioning between cameras. This is different from GetBlendInfo, which retrieves the blend information of the camera which is being blended INTO.

This new function is used to smoothly blend the camera bone's offset into the current camera mode by multiplying it with the first-person camera's weight.