Updated camera modes to track blend-out weights
Previously, when blending between camera modes, the weight of the target mode was tracked, while the weight of the other camera modes was left at "1.0." Now, blend weights are also tracked when blending OUT of a camera mode.

This fixes some issues caused by rapidly changing camera modes, blending between more than two camera modes, and reverting to a previous camera mode before finishing blending to a new one.

This also allows us to track the weight of specific camera modes, regardless of their lifetime. For example, we can now use a "GetCameraWeightByTag" function to get the current weight of a first-person camera, regardless of whether it's currently blending in or out. Previously, we had to retrieve the top camera's weight and
"guess" whether the first-person camera was blending in or out.