Added custom cursor widgets
Created custom cursor widgets for game use. Every cursor type uses the game's "Default" cursor widget. This is because our interface design makes it clear when a button or interactable widget is hovered, clicked, or held, so a custom context-specific cursor is not needed to communicate this.

The only exception is the "Grab Hand Closed" cursor type, which uses a custom cursor to communicate that a clicked widget is intended to be dragged (e.g. sliders).