Implemented "AddWidgets" game feature action
This action listens for an extension request from the player's HUD. Upon receiving the request, it adds a collection of HUD layouts and slotted widgets to the player's user interface. This will be used to drive game mode-based interfaces and HUDs, define by game mode data. This is equivalent to the old UI system's "User Interface Data."

Currently, there is no custom HUD class sending the extension request, so the action does not yet add any layouts or widgets.