Implemented class outline for inventory system
Created the InventoryList and InventoryComponent classes, which will serve as the basis for the inventory system.

I'm choosing to implement a minimalistic inventory system for this game for the sake of simplicity. It should still have the flexibility to handle anything short of a straight-up survival crafting game. But if we ever desire a more robust system, I've added a link to a document that outlines what this could look like.