Implemented ItemDefinition and InventoryItemTraitBase classes
ItemDefinition will be subclassed to actually define items, which will be instantiated in-game into ItemInstances. Item definitions are comprised of a collection of item traits, which define the item's properties. Instead of subclassing, modifying, or expanding the ItemDefinition class, new item properties can be added by creating a new item trait subclass.