Implemented the "inventory utility" editor utility widget
This displays a list of the items in a selected actor's inventory. It also provides buttons to remove, drop, and add inventory items in-editor during runtime.

There are currently two features missing: (1) real-time inventory updates, as the "inventory change message" message broadcasting system has not yet been implemented, and (2) the ability to drop items, as the "droppable" trait has not yet been implemented.