Refactored inventory and equipment to be controller-based
Changed the InventoryComponent to be a controller component, so it can be used on both players and AI, without worrying about where the inventory is stored (player state vs pawn).

This has decoupled the inventory from the player state, meaning it can no longer access information from the pawn data. Because of this, pawns' default items now must be added manually in blueprints (e.g. in OnPossessed).

I'm not sure if I like this new method. It's definitely an improvement over the previous system (which just didn't work that well), but I don't really like making the inventory and equipment modular systems—rather than being built into the game framework like the ability system. I may refactor this again soon.