Fixed first-frame animation update bugs
There were two bugs related to the first update of the base character animation instances. The first bug was caused by animation updates being called before the character movement component's movement was was set. When the movement mode is unset, GetMaxSpeed reteurns 0.0. In our velocity update function, we normalize a number with [0.0, GetMaxSpeed]. Since GetMaxSpeed was returning 0.0 for the first few updates, our character's perceived velocity was its maximum speed (any number normalized from 0 -> 0 is 1). A conditional statement has been added to ensure the character movement component and its movement mode are both initialized before any updates can occur.

The second bug was caused by character transform values being set on the first frame. For example, we initialize our character's aim rotation to (0, 0, 0). But when the player's character spawns, the player's aim could be (180, 90, 0), which would result in an extremely high RotationDelta for the first frame. This was fixed by adding a "bFirstUpdate" variable to the base character animation instance, which gets set to "false" after the first successful update. Any delta-based values now check if bFirstUpdate is true when they get set. If it is, they'll stay at 0.0 for the first update instead of calculating the actual delta.