Removed inventory component from player state
My first iteration of the inventory system placed the inventory component in the player state. When I changed the system to use a controller component instead, Unreal just... didn't complain when I was still adding a controller component to a player state. I removed it, since it wasn't ever supposed to be there.