Version 0.3.4
This version implements the "Inventory" and "Equipment" systems.

An "inventory" is a persistent collection of items owned by player controllers. It is fully data-driven, based around "item definitions," which modularly define the attributes of an item,

"Equipment" is an object that can be "equipped" by a pawn, granting unique on-equip properties and behaviors (such as granting equipment-specific abilities), and usually spawning actors to represent that equipment (such as a sword). With the current system, only one equipment can be equipped at any time. Equipment only exists while equipped. When an object is unequipped, it is always destroyed.

Inventory and equipment is fully replicated. Equipment is integrated with the dual-camera (first- and third-person) camera system.

An "inventory utility" widget has been implemented to aid in testing and debugging inventory and equipment logic.