Implemented the TimedEquipmentNiagaraEffect anim notify state
This is a subclass of the AnimNotifyState_TimedNiagaraEffect anim notify class, which adds support for perspective-based visibility. When this effect is spawned, its visibility will be synced with that of its parent mesh. Because anim notify effects are spawned as attached components, its visibility will automatically be updated when its owning character's perspective changes, as this will propagate the perspective change to all attached components.