Implemented "Death" ability trigger
The "Death" ability (any that the player can activate; usually granted by the GM) is now activated when the player runs out of health. I tested this by adding a new "Death" camera mode that the "Death" ability switches to when activated.

Also fixed some redirector issues with the tags renamed in a previous commit.