Added game mode loading dependency to pawn extension component
The pawn extension component now requires the game mode data to be loaded before it can be progressed to Initializing. Some components require the game mode data (or objects added by it) to properly initialize. For example, the health component requires a health attribute set, which is added by the game mode.

This dependency only affects bots placed in the level, since players aren't spawned until the game mode is loaded.