Implemented procedural kill volume damage
Previously, characters were killed when going out of bounds by having a volume in the world that applied a damage effect when the player left it. This required a volume like this to be manually added for each level, and also inadvertently placed level bounds at the top and sides of each level.

A new system has been added to dynamically kill the player using Unreal's built-in "fall out of world" system. This requires actors to override the "fell out of world" function. This is a much cleaner solution that makes use of Unreal's built-in methods and allows for more precise control of what happens when actors leave the level's bounds.

To support the implementation of a dynamic damage effect, I've added new data and methods to this project's asset manager and global game data classes.

If we ever want to add a system similar to the previous one, where something other than death happens when a player exits the bounds of a level, we can still do so using the ability system volume; we just have to make sure that its bounds lay above the world's kill Z height.