Implemented death camera for environmental death
Created a new base camera type: StaticCameraBase. This camera defines a location and rotation where the camera will be fixed; the rotation can be overridden using the player's controller rotation, allowing them to turn on their own, if desired.

This is now used for environmental deaths—i.e. falling out of the world. It would be pretty boring if we made the player stare into the abyss when they died, so now we place them into a view overlooking the map if they die from falling out of the world.