Created new "TaggedWidget" class
This is a new user widget class that defines a collection of gameplay tags. The widget listens to its owning player's ASC, and automatically hides or reveals itself when its tags are added or removed. This is useful for generically controlling the behavior of widgets without needing a direct reference or search. It is being created to allow us to hide the reticle widget while the player is dead.

Widget is not yet fully functional. This commit is for switching workspaces.