Implemented the "Reset" ability
This is a gameplay ability that can be activated by a "Reset" gameplay event to instantly reset the target player. This would be useful for something like round-based game mode, where players need to be respawned without being killed.

There is currently a bug that is causing minor conflicts with the auto-respawn ability's UI that is unresolved.