Implemented outline for the AbilityWidgetBase class
This is a widget class that will receive events from its owning player's ability system and gameplay abilities. It will be used to give designers extremely easy access to any kind of ability-related events they could want to know about in user widgets. It will be especially helpful to reduce boilerplate code needed for two widgets that make use of the ability system: the ability HUD (which displays core abilities and equipment) and the ultimate ability widget (which only displays the ultimate ability).