Implemented reusable ability cooldowns
Instead of having to use a unique gameplay effect for every single cooldown in the game (as GAS does out-of-the-box), all abilities can now re-use a single global cooldown GE, and override its duration with their own duration (defined in class defaults).

This new feature has been implemented and tested on the Knight's "Joust" and "Warhammer" abilities.