Implemented cooldown start prediction
Abilities can now predict when their cooldowns are applied. This is a built-in feature of GAS—IF you apply the cooldown within the prediction window of the ability's activation. Most of our abilities, however, don't apply their cooldowns until they latently finish (e.g. at the end of a dash). I had to override the default ApplyCooldown function to create a new prediction window just for the cooldown. Now, cooldowns can be predictively applied and rolled-back.