Adjusted how cooldowns are predicted
Adjusted cooldown prediction logic to create a more intuitive user-facing experience. Cooldowns were previously predicted locally, along with their effects (e.g. gameplay messages used to trigger UI responses). Now, cooldowns are still predicted, but their effects are not. This allows clients to always keep up-to-date with the state of their abilities (allowing UI to check for ability logic, like whether the ability is on a cooldown), while keeping the cooldown's effects in sync with the server (meaning the UI's timer will always be accurate).