Implemented the InstigatorEffectsGameplayEffectComponent class
This is a new gameplay effect component that behaves exactly like the "Add Additional Effects" component, except it applies the effects to the INSTIGATOR of the outer effect, instead of the target. This was added to support the modification of the "UltimateCharge" attribute whenever an actor applied the "Damage" effect to another actor.

Also created the "Ultimate" attribute set, which defines data for tracking ultimate abilities. It has not yet implemented.