Implemented UltimateAttributeSet
This is an attribute set that contains an "UltimateCharge" attribute. This attribute is used as a cost for all ultimate abilities. It is automatically incremented for every point of damage and healing dealt, and can be externally modified for things like ultimate charge pick-ups.

Abilities can create and set a cost gameplay effect to define the amount of charge required for their activation. This is a means of balancing the strength of ultimate abilities.

Also removed the InstigatorEffect gameplay effect component, as the means of granting ultimate charge has been updated to not use it. This GE component did not take into account whether the executions of its outer GE were successfully applied, meaning we could grant ultimate charge without actually dealing any damage or healing.