Implemented meta health attribute clamps
Damage and healing executions are now clamped to what can actually be applied. E.g. if we try to deal 200 damage to a target with 50 health, we'll now only deal 50 damage instead of 200. This is because we never want to deal more damage or healing than can actually be applied, because our other systems only care about applied effects. E.g. if we let our execution deal 200 damage, even though to target will only take 50, we'll still grant 200 ultimate charge, when we really want to grant 50.